[mk][top] Use same init categories

+++ mk/top/init.c
@@ -25,7 +25,8 @@ extern const struct lk_init_struct __start_lk_init[];
 extern const struct lk_init_struct __stop_lk_init[];
 
 void lk_init_level(enum lk_init_flags required_flag, uint start_level, uint stop_level) {
-  LTRACEF("flags %#x, start_level %#x, stop_level %#x\n", required_flag, start_level, stop_level);
+  LTRACEF("flags %#x, start_level %#x, stop_level %#x\n", (uint)required_flag, start_level,
+          stop_level);
 
   ASSERT(start_level > 0);
   uint last_called_level = start_level - 1;
